refactor(class): 독립 모듈화를 위한 새로운 클래스 생성 및 분리
